Auth and Session services, Basic mock login
The Auth and Session services were created to handle users logging in
and storing their sessions. Additionally, a Auth-Guard was also created
to keep anonymous users from viewing the account routers without logging
in.

The login page was also updated with a method that calls the Auth
Service's login functionality.