Major update to pricing plans, price boxes, and payment pages
The way the plan data is obtained has changed. In this commit, there is
not hard coded plan data anymore. All plan data must be pulled from the
server on the pricing, dashboard, and upgrade pages. The priceboxes,
paypal, and bitpay component were all updated to match how the data is
pulled asynchronously. The backend service was also updated because the
pricing/plans api is now a get call with a path param for referralCode.
The express server code for both dev and aot were updated to reflect the
change in http method. The plans and session service were also updated
because the call to updated the plan visibility was moved into the
priceboxes component.

Finally, some modules were moved around in the app.module.ts file to
account for new routes being added to the dashboard page. This is part
of a partial update to the new Dashboard design.